Optimize air/water/sound button images + Componentize city gauge
